Refactor classes with implementations and improved naming consistency
The commit reflects major refactoring across server-side code. Important code structures like services and facades have been refactored into interfaces and default implementations. This improves the overall code structure by separating contract and implementation. Furthermore, the commit also includes renaming of several classes and methods for better consistency across the codebase. Throughout the changes, all necessary adjustments have been made in corresponding test classes to align with the refactorings.
