+++ server/src/main/java/org/hkurh/doky/search/solr/DocumentIndexBean.kt
@@ -13,7 +13,4 @@ class DocumentIndexBean {
     var description: String? = null
 }
 
-class DocumentResultBean {
-    @Field
-    var id: String? = null
-}
+
