+++ server/src/main/java/org/hkurh/doky/search/solr/DocumentResultBean.kt
@@ -1,0 +1,8 @@
+package org.hkurh.doky.search.solr
+
+import org.apache.solr.client.solrj.beans.Field
+
+class DocumentResultBean {
+    @Field
+    var id: String? = null
+}
