+++ server/src/main/java/org/hkurh/doky/users/api/UserController.kt
@@ -1,8 +1,5 @@
 package org.hkurh.doky.users.api
 
-import jakarta.validation.constraints.Email
-import jakarta.validation.constraints.Pattern
-import jakarta.validation.constraints.Size
 import org.hkurh.doky.security.DokyAuthority
 import org.hkurh.doky.users.UserFacade
 import org.springframework.http.ResponseEntity
@@ -28,13 +25,3 @@ class UserController(private val userFacade: UserFacade) : UserApi {
         return ResponseEntity.noContent().build<Any>()
     }
 }
-
-data class UserDto (
-    var id: Long,
-    @Email
-    var uid: String,
-    var name: String? = null,
-    @Size(min = 8, max = 32, message = "Length should be from 8 to 32 characters")
-    @Pattern(regexp = "^[a-zA-Z\\d!@#$%^&*()_\\-+]*$")
-    var password: String? = null
-)
