+++ server/src/main/java/org/hkurh/doky/users/api/UserDto.kt
@@ -1,0 +1,15 @@
+package org.hkurh.doky.users.api
+
+import jakarta.validation.constraints.Email
+import jakarta.validation.constraints.Pattern
+import jakarta.validation.constraints.Size
+
+data class UserDto(
+    var id: Long,
+    @Email
+    var uid: String,
+    var name: String? = null,
+    @Size(min = 8, max = 32, message = "Length should be from 8 to 32 characters")
+    @Pattern(regexp = "^[a-zA-Z\\d!@#$%^&*()_\\-+]*$")
+    var password: String? = null
+)
