+++ server/src/test/java/org/hkurh/doky/authorization/AuthorizationUserControllerTest.kt
@@ -5,18 +5,15 @@ import org.hkurh.doky.users.UserFacade
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.mockito.InjectMocks
-import org.mockito.Spy
 import org.mockito.kotlin.mock
 
 
 @DisplayName("AuthorizationUserController unit test")
 class AuthorizationUserControllerTest : DokyUnitTest {
 
-    @InjectMocks
-    @Spy
-    lateinit var controller: AuthorizationUserController
+
     private val userFacade: UserFacade = mock()
+    private var controller = AuthorizationUserController(userFacade)
 
     @Test
     @DisplayName("Should generate token when login")
